Backend/servicios-IA-voz/texto-voz-api-ms/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-multipart>=0.0.6
aiofiles>=23.1.0
